requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
qrcode==8.2
cairosvg==2.7.1
Pillow==11.1.0